Fix XR Plug-in Management Error
During development in Unity 2023.2.6f1, I encountered an error related to XR Plug-in Management version 4.4.1. The error message "error CS0117: ‘BuildTargetGroup’ does not contain a definition for ‘VisionOS’" appeared when trying to install this version.

After investigating, it was found that this error was due to a bug in the Unity package itself. Unity introduced support for VisionOS in version 4.4.1, but it caused issues, especially for projects using Unity 2023.2.6f1 on macOS.

To resolve this problem, I downgraded the XR Plug-in Management package to version 4.4.0 by editing the Packages/manifest.json file and changing the package version:

"com.unity.xr.management": "4.4.0"

This downgrade resolved the error and allowed me to continue working on the project without issues.

Note: Unity may release updates or fixes for this package in the future, so it's a good idea to check for updates and test compatibility with your specific Unity version.